perf: Implement faster BitArray/BitMatrix data representation
Compressing a greyscale (8bit per pixel) image into a compressed (1bit
per pixel) image is obviously more space efficient but it is
unfortunately less computation/data-access efficient (on standard intel
hardware).

This changeset implements an alternative byte-per-pixel version of
BitArray and BitMatrix. This improves the runtime for the falsepositives
tests (on said intel hardware) by roughly 20%. The increased memory
usage is non-observable for the ReaderTest application.

This can be switched on and off by the `#define ZX_FAST_BIT_STORAGE` in
`ZXConfig.h`. It might make sense to make this configuable from within
CMake.

+++ core/src/oned/ODRSS14Reader.cpp
@@ -108,8 +108,8 @@ ParseFoundFinderPattern(const BitArray& row, int rowNumber, bool right, BitArray
 	}
 
 	return {value,
-			range.begin - row.begin(),
-			range.end - row.begin(),
+			static_cast<int>(range.begin - row.begin()),
+			static_cast<int>(range.end - row.begin()),
 			{ResultPoint(start, rowNumber), ResultPoint(end, rowNumber)}};
 }
 
@@ -368,6 +368,7 @@ AddOrTally(std::list<RSS::Pair>& possiblePairs, const RSS::Pair& pair)
 			return;
 		}
 	}
+//	printf("found new pair\n"); fflush(stdout);
 	possiblePairs.push_back(pair);
 }
